attempt No.2 to fix ScannerError
[Python and PYAML - yaml.scanner.ScannerError: mapping values are not allowed here - Stack Overflow](https://stackoverflow.com/questions/9055371/python-and-pyaml-yaml-scanner-scannererror-mapping-values-are-not-allowed-her)

For anyone who comes here and finds that even if they have spaces after the colon, they still get this error

You can also get this error if you copy the yaml text from some formatted source (for me it was a Slack message). This will invisibly swap in non-ASCII characters that the standard YAML reader can't read, but which look the same.

Solution is to only copy from raw, ASCII source.